refactor(utils): improve logging and update related imports
