+++ packages/utils/package.json
@@ -6,6 +6,7 @@
   "sideEffects": false,
   "exports": {
     ".": "./src/index.ts",
+    "./logging": "./src/lib/logging/index.ts",
     "./package.json": "./package.json"
   },
   "nx": {
