+++ packages/utils/src/lib/logging/index.ts
@@ -1,0 +1,2 @@
+export { getLogger } from './get-logger'
+export type * from './types'
